+++ src/django_http_inspector/__init__.py
@@ -3,4 +3,4 @@
 from .wrapper.wsgi import InspectorWSGI
 
 __all__ = ["InspectorWSGI"]
-__version__ = "0.1.1"
+__version__ = "0.1.2"